[Test] Add pytest skeleton.
Add a pytest function to exercise the resolver.
Add more expanded logging to the resolver skeleton to facilitate this.
Add a tests folder. The `plugInfo.json` lives there, as the file is
really only for our internal tests/as an example, real use cases will
have folks roll their own, since the paths are application dependent.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>