[Build] Integrate OpenAssetIO
(#15)

Add OpenAssetIO as a dependency of the resolver, so that OpenAssetIO
methods can be use to implement resolver interface methods.

Add OpenAssetIO building to the test automation along with this.
Building OpenAssetIO is done locally, as unlike a lot of our other
projects, USD is a C++ host, so can't just grab the thing from PyPI.
This is slated to change in the future when conan lights up.

The workflow for building the USD resolver is, at the moment, simply
"have openassetio installed on your system and reference it." there is
no formal dependency management.

We've used our own `openassetio-build` image to build OpenAssetIO on CI,
as this sidesteps a lot of issues with having to get dependencies.
As this is a compatible container with `asfw/vfxall`, this allows us
to transfer the build over to that container in order to build and run
the resolver itself, as the resolver requires the very chunky USD
install, which isn't on the `openassetio-build` image.

There's a fair amount of environment variables that need to be set to
have OpenAssetIO and USD both playing happy here. In many ways there's
not much we can do to avoid this, but hopefully having a conan C++
OpenAssetIO build will allow us to at least remove the issue of having
to build OpenAssetIO manually for applications like this one.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>