[Tests] Use a single BAL library across all tests
The USD resolver plugin is only initialised once for the whole suite of
tests and we have no way to poke BAL to force a reinitialision with a
different library.

So combine all per-test `bal_libary.json` into a single library.

Add missing env var fixtures in the resolver `pytest` tests.

Signed-off-by: David Feltell <david.feltell@foundry.com>