[Docs] Improve readme documentation
The README documentation had gotten out of date and needed updating.
Furthermore, as the resolver is going to be left in a working, but quite
prototypey phase, it is important to note the project state, so people
aren't unpleasantly shocked when the resolver doesen't do everything out
of the box.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>