[Core] Move resolution to _Resolve
When we made this PoC, we noted that in order to support querying
extension, mtime, from the manager, as well as to support publishing
we'd need access to the entity reference in `_OpenAssetForWrite` and
other methods.

We opted to pass the ref through `_Resolve` so it would be available to
these other methods, and actually do the resolve in `_OpenAsset` and
friends.

We know performance is important, and we need to ensure that early
adopters don't see significantly increased overhead due to OpenAssetIO.

Looking again, we can split behaviour, so that the old approach is only
done in `_ResolveForNewAsset`. This means that the common case (read)
won't end up with numerous redundant resolver per read.

Signed-off-by: Tom Cowland <tom@foundry.com>