[Core] Remove debug logging
We put this in when we were learning about when and where USD invoked
the various methods of the API. People have indicated that the overhead
of the wrapper to OpenAssetIO would be a key factor during assessment.

This commit removes this logging, to avoid any unnecessary overhead.

It also makes the code somewhat easier to grok, which should help people
using this as a reference implementation.

Signed-off-by: Tom Cowland <tom@foundry.com>