[Core] Support C++ managers
Part of #40. Add support for C++ manager plugins via use of the hybrid
plugin system, which additionally means that Python and C++ plugins can
be composed.

Make the tests a bit more agnostic to the particular manager. Add a
config file suitable for using SimpleCppManager as the plugin used to
test against. Modify test for error behaviour when given an invalid
asset reference so that it works for both BAL and SimpleCppManager -
i.e. instead of `bal:///` ("maformedEntityReference" for BAL only) use
`bal:///doesntexist`, which causes an "entityResolutionError" in both
BAL and SimpleCppManager.

Signed-off-by: David Feltell <david.feltell@foundry.com>