[Tests] Allow overriding OPENASSETIO_DEFAULT_CONFIG
Part of #40. In preparation for running the tests through another
manager (i.e. SimpleCppManager), do not overwrite the
value of the `OPENASSETIO_DEFAULT_CONFIG` environment variable, so that
the manager to be used by the test can be configured externally.

This does add yet another variable that must be set before the tests can
run and pass, but there are already several others that are required
anyway (`LD_LIBRARY_PATH`, `OPENASSETIO_PLUGIN_PATH`,
`PXR_PLUGINPATH_NAME`, `PYTHONPATH`) , so it's not that much extra
bother, and forcing the env var to be set externally means we can be
sure the intended manager is used for the tests.

Signed-off-by: David Feltell <david.feltell@foundry.com>