[Core] Use OpenAssetIO context from manager
The Context instance is passed into every API method. The Context may
bundle arbitrary state from the manager. In order for the Context to be
created with the appropriate state by the manager, the host is required
to construct a Context object using the `Manager.createContext()`
method. However, we were not doing that and instead constructed the
Context directly, which should never be done (other than for trivial
cases e.g. unit tests).

Signed-off-by: David Feltell <david.feltell@foundry.com>